feat: Sections.toStations 정렬 기능 테스트 추가 및 구현
