+++ src/main/java/subway/application/StationService.java
@@ -1,8 +1,6 @@
 package subway.application;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import subway.dao.SectionDao;
@@ -51,57 +49,9 @@ public class StationService {
 
     public List<Station> findStationByLineId(Long lineId) {
         List<Section> sectionList = sectionDao.findAllByLineId(lineId);
-
         Sections sections = new Sections(sectionList);
-//
-//        List<Section> result = sort(sections);
-//        List<Station> stationsa = new ArrayList<>();
-//
-//        Stations stations = sections.toStations();
-//
-//        stations.add(result.get(0).getUpStation());
-//        stations.addAll(result.stream().map(Section::getDownStation).collect(Collectors.toList()));
-
         return sections.toStations();
     }
 
-    private List<Section> sort(List<Section> sections) {
-        Section pivot = getFirstSection(sections);
-        return getSortedSections(sections, pivot);
-    }
 
-    private Section getFirstSection(List<Section> sections) {
-        Section pivot = sections.get(0);
-        while (true) {
-            Optional<Section> temp = findUpSection(sections, pivot);
-            if (temp.isEmpty()) {
-                return pivot;
-            }
-            pivot = temp.get();
-        }
-    }
-
-    private List<Section> getSortedSections(List<Section> sections, Section pivot) {
-        List<Section> result = new ArrayList<>();
-        while (true) {
-            result.add(pivot);
-            Optional<Section> temp = findDownSection(sections, pivot);
-            if (temp.isEmpty()) {
-                return result;
-            }
-            pivot = temp.get();
-        }
-    }
-
-    private Optional<Section> findUpSection(List<Section> sections, Section pivot) {
-        return sections.stream()
-                .filter(section -> section.getDownStationId().equals(pivot.getUpStationId()))
-                .findAny();
-    }
-
-    private Optional<Section> findDownSection(List<Section> sections, Section pivot) {
-        return sections.stream()
-                .filter(section -> section.getUpStationId().equals(pivot.getDownStationId()))
-                .findAny();
-    }
 }
