+++ src/main/java/subway/domain/Sections.java
@@ -3,13 +3,69 @@ package subway.domain;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 public class Sections {
 
     private List<Section> sections;
 
     public Sections(List<Section> sections) {
-        this.sections = sections;
+        this.sections = sort(sections);
+    }
+
+    public List<Station> toStations() {
+        if (sections.isEmpty()) {
+            return List.of();
+        }
+        List<Station> result = sections.stream()
+                .map(Section::getUpStation)
+                .collect(Collectors.toList());
+        result.add(sections.get(sections.size() - 1).getDownStation());
+        return result;
+    }
+
+    private List<Section> sort(List<Section> sections) {
+        if (sections.isEmpty()) {
+            return List.of();
+        }
+        Section pivot = getFirstSection(sections);
+        return getSortedSections(sections, pivot);
+    }
+
+    private Section getFirstSection(List<Section> sections) {
+        Section pivot = sections.get(0);
+        while (true) {
+            Optional<Section> temp = findUpSection(sections, pivot);
+            if (temp.isEmpty()) {
+                return pivot;
+            }
+            pivot = temp.get();
+        }
+    }
+
+    private List<Section> getSortedSections(List<Section> sections, Section pivot) {
+        List<Section> result = new ArrayList<>();
+        while (true) {
+            result.add(pivot);
+            Optional<Section> temp = findDownSection(sections, pivot);
+            if (temp.isEmpty()) {
+                return result;
+            }
+            pivot = temp.get();
+        }
+    }
+
+    private Optional<Section> findUpSection(List<Section> sections, Section pivot) {
+        return sections.stream()
+                .filter(section -> section.getDownStationId().equals(pivot.getUpStationId()))
+                .findAny();
+    }
+
+    private Optional<Section> findDownSection(List<Section> sections, Section pivot) {
+        return sections.stream()
+                .filter(section -> section.getUpStationId().equals(pivot.getDownStationId()))
+                .findAny();
     }
 
     public List<Section> getSections() {
@@ -38,9 +94,5 @@ public class Sections {
         return "Sections{" +
                 "sections=" + sections +
                 '}';
-    }
-
-    public List<Station> toStations() {
-        return List.of();
     }
 }
